refactor: rename Progress Bar demo files
