+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/00-progress-bar-docs.twig
@@ -6,6 +6,3 @@
   componentName: "progress-bar",
   usage: usage
 } only %}
-
-{# not passing in `only` here since the global link variable needs to get passed in #}
-{% include "@pl/_progress-bar-demos.twig" %}
