+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/10-basic-usage/10-progress-bar--basic-usage.twig
