+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/10-basic-usage/value-0/_progress-bar--basic-usage--value-0.twig
