+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/10-basic-usage/value-50/_progress-bar--basic-usage--value-50.twig
