+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/15-theming/15-progress-bar--theming.twig
@@ -1,0 +1,13 @@
+{% set schema = bolt.data.components["@bolt-components-button"].schema %}
+{% set themes = get_data("@bolt/themes.schema.json") %}
+
+<div class="u-bolt-shadow-level-10">
+  {% for theme in themes.properties.colors.enum %}
+    <div class="u-bolt-padding-medium t-bolt-{{ theme }}">
+      <h3>{{ theme }} theme</h3>
+      {% include "@bolt-components-progress-bar/progress-bar.twig" with {
+        value: "50",
+      } only %}
+    </div>
+  {% endfor %}
+</div>
