+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/15-theming/16-progress-bar--theming--toolbar-example.twig
