+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/20-min-max/20-progress-bar--min-max.twig
