+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/25-striped-and-animated/26-progress-bar--animated.twig
