+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/30-title/30-progress-bar--title.twig
