+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/30-title/31-progress-bar--title--with-value-position-top.twig
