+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/35-value-position/35-progress-bar--value-position.twig
@@ -10,7 +10,7 @@
   value: "75",
 } only %}
 
-<h2>valuePosition: <code>outside</code></h2>
+<h2>valuePosition: <code>outside</code> (default)</h2>
 {% include "@bolt-components-progress-bar/progress-bar.twig" with {
   valuePosition: "outside",
   value: "75",
