+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/40-advanced-usage--stepper/-progress-bar--stepper-example.js
