+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/40-advanced-usage--stepper/40-progress-bar--stepper-example.twig
