+++ docs-site/src/pages/pattern-lab/_patterns/02-components/progress-bar/45-advanced-usage--programmatically-updated/45-progress-bar--programmatically-updated-example.twig
